// backend/src/models/kanban/JiraBlockReasonEnum.ts
export enum JiraBlockReasonEnum {
  DEPENDENCIES_NOT_WORK = "dependencies_not_work",
  SIT_ENV_DOWN = "sit_env_down",
  PRIORITY_CHANGE = "priority_change",
  SOLUTION_REVIEW = "solution_review",
  PR_REVIEW = "pr_review",
  QUESTION_TO_BE_ANSWERED = "question_to_be_answered",
  TAKE_LEAVE = "take_leave",
  OTHERS = "others",
}